Fix non-standard writing
